requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bidict==0.23.1
bitarray==3.7.2
certifi==2025.10.5
charset-normalizer==3.4.3
ckzg==2.1.5
click==8.3.0
cytoolz==1.0.1
distro==1.9.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.118.3
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.7.0
numpy==2.3.3
openai==2.3.0
pandas==2.3.3
parsimonious==0.10.0
propcache==0.4.1
pycryptodome==3.23.0
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.1
pytz==2025.2
pyunormalize==17.0.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
rpds-py==0.27.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tiktoken==0.12.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
virtuals-acp==0.2.14
web3==7.13.0
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.22.0
